+++ backend/controllers/student.js
@@ -5,15 +5,29 @@ module.exports.getProfileData = async (req, res) => {
   const { id } = req.params;
   const user = await User.findOne({ where: { id: id } });
   if (user == undefined) {
-    res.send({ err: "No such account" });
+    res.send({ status: "error", data: null, err: "No such account" });
   } else {
-    const student = await Student.findOrCreate({
-      where: { userId: id },
-      default: { userId: id },
-    });
-    const dataRequired =
-      student[0].rollno === -1 || student[0].prn === "required" || student[1];
-    res.send({ isFirstLogin: dataRequired, data: student[0] });
+    let { role } = user;
+    if (role !== "student") {
+      res.send({
+        status: "fail",
+        data: null,
+        err: `user: ${id} is not a student`,
+      });
+    } else {
+      const student = await Student.findOrCreate({
+        where: { userId: id },
+        default: { userId: id },
+        include: { model: User, attributes: ["email", "fullname"] },
+      });
+      const dataRequired =
+        student[0].rollno === -1 || student[0].prn === "required" || student[1];
+      res.send({
+        status: "success",
+        data: { isFirstLogin: dataRequired, student: student[0] },
+        err: null,
+      });
+    }
   }
 };
 
@@ -24,8 +38,15 @@ module.exports.updateProfileData = async (req, res) => {
       id: req.params.id,
     },
   });
+  if (!role) {
+    res.send({ status: "error", data: null, err: "No such account" });
+  }
   if (role !== "student") {
-    res.status(403).send(`userId: ${req.params.id} is not a student.`);
+    res.status(403).send({
+      status: "fail",
+      data: null,
+      err: `userId: ${req.params.id} is not a student.`,
+    });
   } else {
     const result = await Student.update(
       { ...req.body.student },
@@ -36,11 +57,21 @@ module.exports.updateProfileData = async (req, res) => {
 };
 
 module.exports.deleteStudent = async (req, res) => {
+  const exists = await Student.findOne({
+    where: { userId: req.params.id },
+  });
+  if (!exists) {
+    res.send({ status: "error", data: null, err: "No such account" });
+    return;
+  }
   const result = await Student.destroy({
     where: { userId: req.params.id },
   });
-  if (result === 1) res.send("Deleted");
-  else res.status(500).send("Something went wrong");
+  if (result === 1) res.send({ status: "success", data: null, err: null });
+  else
+    res
+      .status(500)
+      .send({ status: "error", data: null, err: "Something went wrong" });
 };
 
 module.exports.search = async (req, res) => {
@@ -56,5 +87,5 @@ module.exports.search = async (req, res) => {
       attributes: ["fullname"],
     },
   });
-  res.send(students);
+  res.send({ status: "success", data: students, err: null });
 };
